python_app/requirements.txt
matplotlib
python-dateutil
numpy
scikit-fuzzy
